Refactor component structure: move Card and Tooltip components to common directory for better organization and maintainability
